+++ CHANGELOG.md
@@ -4,6 +4,10 @@ The following is a curated list of changes in the Enact sandstone module, newest
 
 ## [unreleased]
 
+### Added
+
+- `sandstone/Scroller` prop `scrollbarTrackCss` to customize scroll track and thumb style
+
 ### Fixed
 
 - `sandstone/VirtualList` to not block key down events after panel transition